Remove and fix article-related tests
These tests are better performed by DCR itself now that we are
close to 100% of vanilla articles migrated:

* avoids coupling our build process to DCR availability
* avoids duplication of tests between DCR+Frontend
* takes us closer to the slimline Frontend we want

This is a drastic move, but an integration test with DCR without
network dependencies is not sensible; DCR is too complex to stub.
