+++ article/test/ArticleControllerTest.scala
@@ -1,37 +1,11 @@
 package test
 
-import com.gu.contentapi.client.model.v1.Blocks
 import controllers.ArticleController
-import model.Cached.RevalidatableResult
-import model.dotcomrendering.PageType
-import model.{ApplicationContext, Cached, PageWithStoryPackage}
 import org.apache.commons.codec.digest.DigestUtils
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{BeforeAndAfterAll, DoNotDiscover, FlatSpec, Matchers}
-import play.api.libs.ws.WSClient
-import play.api.mvc.{RequestHeader, Result}
 import play.api.test.Helpers._
 import play.api.test._
-import play.twirl.api.Html
-
-import scala.collection.JavaConverters._
-import scala.concurrent.{ExecutionContext, Future}
-
-// I had trouble getting Mockito to play nicely with how scala is using implicits and consts so I've introduced
-// these
-
-class FakeRemoteRender(implicit context: ApplicationContext) extends renderers.DotcomRenderingService {
-  override def getArticle(
-      ws: WSClient,
-      path: String,
-      article: PageWithStoryPackage,
-      blocks: Blocks,
-      pageType: PageType,
-  )(implicit request: RequestHeader): Future[Result] = {
-    implicit val ec = ExecutionContext.global
-    Future(Cached(article)(RevalidatableResult.Ok(Html("OK"))))
-  }
-}
 
 @DoNotDiscover class ArticleControllerTest
     extends FlatSpec
@@ -45,37 +19,18 @@ class FakeRemoteRender(implicit context: ApplicationContext) extends renderers.D
     with WithTestContentApiClient {
 
   val articleUrl = "environment/2012/feb/22/capitalise-low-carbon-future"
-  val guuiArticle = "world/2018/sep/13/give-pizza-a-chance-south-koreans-pa-weight-to-thwart-conscription"
   val liveBlogUrl = "global/middle-east-live/2013/sep/09/syria-crisis-russia-kerry-us-live"
-  val sudokuUrl = "lifeandstyle/2013/sep/09/sudoku-2599-easy"
 
   lazy val articleController = new ArticleController(
     testContentApiClient,
     play.api.test.Helpers.stubControllerComponents(),
     wsClient,
-  )
-
-  lazy val guuiController = new ArticleController(
-    testContentApiClient,
-    play.api.test.Helpers.stubControllerComponents(),
-    wsClient,
-    new FakeRemoteRender(),
+    new DCRFake(),
   )
 
   "Article Controller" should "200 when content type is article" in {
     val result = articleController.renderArticle(articleUrl)(TestRequest(articleUrl))
     status(result) should be(200)
-  }
-
-  "Article Controller" should "200 for guui articles" in {
-    val result = guuiController.renderArticle(guuiArticle)(TestRequest(guuiArticle))
-    status(result) should be(200)
-  }
-
-  it should "return article headline" in {
-    val result = articleController.renderHeadline(articleUrl)(TestRequest(articleUrl))
-    status(result) should be(200)
-    contentAsString(result) shouldBe "We must capitalise on a low-carbon future | Norman Baker"
   }
 
   it should "200 when content type is live blog" in {
@@ -88,11 +43,6 @@ class FakeRemoteRender(implicit context: ApplicationContext) extends renderers.D
     val body = contentAsString(result)
     body should include(""""inBodyInternalLinkCount":38""")
     body should include(""""inBodyExternalLinkCount":42""")
-  }
-
-  it should "200 when content type is sudoku" in {
-    val result = articleController.renderArticle(sudokuUrl)(TestRequest(sudokuUrl))
-    status(result) should be(200)
   }
 
   it should "not cache 404s" in {
@@ -127,8 +77,6 @@ class FakeRemoteRender(implicit context: ApplicationContext) extends renderers.D
     )
   }
 
-  val expiredArticle = "football/2012/sep/14/zlatan-ibrahimovic-paris-st-germain-toulouse"
-
   it should "know which backend served the request" in {
     val result = route(
       app,
@@ -145,21 +93,4 @@ class FakeRemoteRender(implicit context: ApplicationContext) extends renderers.D
     val result = route(app, TestRequest("/stage/2015/jul/15/alex-edelman-steve-martin-edinburgh-fringe?index=7")).head
     header("Surrogate-Key", result).head should be(expectedSurrogateKey)
   }
-
-  "International users" should "be in the International edition" in {
-    val request = TestRequest("/world/2014/sep/24/radical-cleric-islamic-state-release-british-hostage-alan-henning")
-      .withHeaders(
-        "X-GU-Edition" -> "int",
-      )
-    val result = route(app, request).head
-    contentAsString(result) should include("\"edition\":\"INT\"")
-  }
-
-  "Interactive articles" should "provide a boot.js script element as a main embed" in goTo(
-    "/sport/2015/sep/11/how-women-in-tennis-achieved-equal-pay-us-open",
-  ) { browser =>
-    import browser._
-    $(".media-primary > .element-interactive").attributes("data-interactive").asScala.head should endWith("boot.js")
-  }
-
 }
