+++ article/test/TestAppLoader.scala
@@ -1,11 +1,17 @@
 import app.FrontendComponents
 import play.api.ApplicationLoader.Context
 import play.api.BuiltInComponentsFromContext
+import renderers.DotcomRenderingService
 import test.WithTestContentApiClient
+import test.DCRFake
 
 trait TestComponents extends WithTestContentApiClient {
   self: AppComponents =>
+
   override lazy val contentApiClient = testContentApiClient
+
+  // Relying on DCR output for tests is always a mistake.
+  override lazy val remoteRender: DotcomRenderingService = new DCRFake()
 }
 
 class TestAppLoader extends AppLoader {
